scripts/requirements.txt
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.3
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pycryptodome==3.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
